@@ -59,6 +59,7 @@
     "babel-eslint": "8.2.6",
     "babel-jest": "23.4.0",
     "babel-loader": "7.1.5",
+    "babel-plugin-dynamic-import-node": "2.0.0",
     "babel-plugin-flow-runtime": "0.17.0",
     "babel-plugin-syntax-dynamic-import": "6.18.0",
     "babel-plugin-transform-async-to-generator": "6.24.1",
@@ -77,8 +78,8 @@
     "babel-register": "6.26.0",
     "babel-runtime": "6.26.0",
     "bundlesize": "0.17.0",
-    "cross-env": "5.1.4",
     "codecov": "3.0.4",
+    "cross-env": "5.1.4",
     "css-loader": "0.28.10",
     "element-react": "1.4.8",
     "element-theme-default": "1.4.13",
@@ -114,8 +115,8 @@
     "ora": "1.4.0",
     "prop-types": "15.6.1",
     "puppeteer": "1.1.1",
-    "react": "16.2.0",
-    "react-dom": "16.2.0",
+    "react": "16.4.2",
+    "react-dom": "16.4.2",
     "react-hot-loader": "4.2.0",
     "react-router-dom": "4.2.2",
     "react-syntax-highlighter": "5.8.0",
