@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {Component, Fragment} from 'react';
 import PropTypes from 'prop-types';
 import {Loading, MessageBox} from 'element-react';
 import isEmpty from 'lodash/isEmpty';
@@ -9,17 +9,17 @@ import API from '../../utils/api';
 import PackageList from '../../components/PackageList';
 import Search from '../../components/Search';
 
-
-export default class Home extends React.Component {
+export default class Home extends Component {
   static propTypes = {
-    children: PropTypes.element
-  }
+    children: PropTypes.element,
+    isUserLoggedIn: PropTypes.bool
+  };
 
   state = {
     loading: true,
     fistTime: true,
     query: ''
-  }
+  };
 
   constructor(props) {
     super(props);
@@ -31,8 +31,7 @@ export default class Home extends React.Component {
     this.loadPackages();
   }
 
-
-  componentDidUpdate(prevProps, prevState) { // eslint-disable-line no-unused-vars
+  componentDidUpdate(prevProps, prevState) {
     if (prevState.query !== this.state.query) {
       if (this.req && this.req.abort) this.req.abort();
       this.setState({
@@ -44,6 +43,10 @@ export default class Home extends React.Component {
       } else {
         this.searchPackage(this.state.query);
       }
+    }
+
+    if (prevProps.isUserLoggedIn !== this.props.isUserLoggedIn) {
+      this.loadPackages();
     }
   }
 
@@ -57,11 +60,11 @@ export default class Home extends React.Component {
           loading: false
         });
       }
-    } catch (err) {
+    } catch (error) {
       MessageBox.msgbox({
         type: 'error',
         title: 'Warning',
-        message: 'Unable to load package list: ' + err.message
+        message: `Unable to load package list: ${error.message}`
       });
     }
   }
@@ -98,11 +101,16 @@ export default class Home extends React.Component {
   }
 
   render() {
+    const {packages, loading} = this.state;
     return (
-      <div>
+      <Fragment>
         {this.renderSearchBar()}
-        {this.state.loading ? this.renderLoading() : this.renderPackageList()}
-      </div>
+        {loading ? (
+          <Loading text="Loading..." />
+        ) : (
+          <PackageList help={isEmpty(packages) === true} packages={packages} />
+        )}
+      </Fragment>
     );
   }
 
@@ -111,13 +119,5 @@ export default class Home extends React.Component {
       return;
     }
     return <Search handleSearchInput={this.handleSearchInput} />;
-  }
-
-  renderLoading() {
-    return <Loading text="Loading..." />;
-  }
-
-  renderPackageList() {
-    return <PackageList help={this.state.fistTime} packages={this.state.packages} />;
   }
 }
