@@ -1,29 +1,46 @@
-import React from 'react';
+import React, {Component} from 'react';
+import PropTypes from 'prop-types';
 import {HashRouter as Router, Route, Switch} from 'react-router-dom';
-import {asyncComponent} from './utils/asyncComponent';
 
-import Header from './components/Header';
-import Footer from './components/Footer';
+import {asyncComponent} from './utils/asyncComponent';
 
 const DetailPackage = asyncComponent(() => import('./modules/detail'));
 const HomePage = asyncComponent(() => import('./modules/home'));
 
-const RouterApp = () => {
-  return (
-    <Router>
-      <div className="page-full-height">
-        <Header/>
+class RouterApp extends Component {
+  static propTypes = {
+    isUserLoggedIn: PropTypes.bool
+  };
+  render() {
+    const {isUserLoggedIn} = this.props;
+    return (
+      <Router>
         <div className="container">
           <Switch>
-            <Route exact path="/(search/:keyword)?" component={ HomePage } />
-            <Route exact path="/detail/@:scope/:package" component={DetailPackage} />
-            <Route exact path="/detail/:package" component={DetailPackage} />
+            <Route
+              exact
+              path="/(search/:keyword)?"
+              render={() => <HomePage isUserLoggedIn={isUserLoggedIn} />}
+            />
+            <Route
+              exact
+              path="/detail/@:scope/:package"
+              render={(props) => (
+                <DetailPackage {...props} isUserLoggedIn={isUserLoggedIn} />
+              )}
+            />
+            <Route
+              exact
+              path="/detail/:package"
+              render={(props) => (
+                <DetailPackage {...props} isUserLoggedIn={isUserLoggedIn} />
+              )}
+            />
           </Switch>
         </div>
-        <Footer/>
-      </div>
-    </Router>
-  );
-};
+      </Router>
+    );
+  }
+}
 
 export default RouterApp;
